requirements.txt
uvicorn[standard]
fastapi==0.70.0
pillow==9.0.1
python-multipart